// apps/web/modules/ee/contacts/api/management/contact-attribute-keys/[contactAttributeKeyId]/types/contact-attribute-keys.ts
import { z } from "zod";

export const ZContactAttributeKeyCreateInput = z.object({
  key: z.string(),
  description: z.string().optional(),
  type: z.enum(["custom"]),
  environmentId: z.string(),
});
export type TContactAttributeKeyCreateInput = z.infer<typeof ZContactAttributeKeyCreateInput>;

export const ZContactAttributeKeyUpdateInput = z.object({
  description: z.string().optional(),
  name: z.string().optional(),
});

export type TContactAttributeKeyUpdateInput = z.infer<typeof ZContactAttributeKeyUpdateInput>;